move plotting functions to plotting.py
