+++ src/nmem/analysis/analysis.py
@@ -1,20 +1,10 @@
 import collections
 import collections.abc
 import os
-from typing import Any, List, Literal, Tuple
+from typing import Any, Literal, Tuple
 
-import matplotlib as mpl
-import matplotlib.font_manager as fm
 import numpy as np
 import scipy.io as sio
-from matplotlib import cm, ticker
-from matplotlib import colors as mcolors
-from matplotlib import pyplot as plt
-from matplotlib.axes import Axes
-from matplotlib.collections import PolyCollection
-from matplotlib.colors import LogNorm, Normalize
-from matplotlib.ticker import LogLocator, MaxNLocator, MultipleLocator, NullFormatter
-from mpl_toolkits.mplot3d import Axes3D
 
 from nmem.calculations.calculations import (
     calculate_heater_power,
@@ -24,184 +14,6 @@ from nmem.measurement.cells import CELLS
 
 SUBSTRATE_TEMP = 1.3
 CRITICAL_TEMP = 12.3
-RBCOLORS = {0: "blue", 1: "blue", 2: "red", 3: "red"}
-
-
-def plot_enable_sweep(
-    ax: plt.Axes,
-    dict_list: list[dict],
-    range=None,
-    add_errorbar=False,
-    add_colorbar=False,
-):
-    N = len(dict_list)
-    if range is not None:
-        dict_list = dict_list[range]
-    # ax, ax2 = plot_enable_write_sweep_multiple(ax, dict_list[0:6])
-    ax = plot_enable_write_sweep_multiple(
-        ax, dict_list, add_errorbar=add_errorbar, N=N, add_colorbar=add_colorbar
-    )
-
-    ax.set_ylabel("BER")
-    ax.set_xlabel("$I_{\mathrm{enable}}$ [$\mu$A]")
-    return ax
-
-
-def plot_enable_sweep_markers(ax: plt.Axes, dict_list: list[dict]):
-    ax.yaxis.set_major_locator(plt.MultipleLocator(0.20))
-    ax.set_ylim([8.3, 9.7])
-
-    write_temp_array = np.empty((len(dict_list), 4))
-    write_current_array = np.empty((len(dict_list), 1))
-    enable_current_array = np.empty((len(dict_list), 4))
-    for j, data_dict in enumerate(dict_list):
-        bit_error_rate = get_bit_error_rate(data_dict)
-        berargs = get_bit_error_rate_args(bit_error_rate)
-        write_current = get_write_current(data_dict)
-        write_temps = get_channel_temperature_sweep(data_dict)
-        enable_currents = get_enable_current_sweep(data_dict)
-        write_current_array[j] = write_current
-        critical_current_zero = get_critical_current_heater_off(data_dict)
-        for i, arg in enumerate(berargs):
-            if arg is not np.nan:
-                write_temp_array[j, i] = write_temps[arg]
-                enable_current_array[j, i] = enable_currents[arg]
-    markers = ["o", "s", "D", "^"]
-    for i in range(4):
-        ax.plot(
-            enable_current_array[:, i],
-            write_current_array,
-            linestyle="--",
-            marker=markers[i],
-            markeredgecolor="k",
-            markeredgewidth=0.5,
-            color=RBCOLORS[i],
-        )
-    ax.set_ylim(0, 100)
-    ax.set_xlim(250, 340)
-    ax.yaxis.set_major_locator(plt.MultipleLocator(20))
-    ax.yaxis.set_minor_locator(plt.MultipleLocator(10))
-    ax.xaxis.set_major_locator(plt.MultipleLocator(25))
-    ax.xaxis.set_minor_locator(plt.MultipleLocator(5))
-    ax.grid()
-    ax.set_ylabel("$I_{\mathrm{write}}$ [$\mu$A]")
-    ax.set_xlabel("$I_{\mathrm{enable}}$ [$\mu$A]")
-    ax.legend(
-        [
-            "$I_{1}$",
-            "$I_{0}$",
-            "$I_{0,\mathrm{inv}}$",
-            "$I_{1,\mathrm{inv}}$",
-        ],
-        loc="upper right",
-        frameon=True,
-        ncol=2,
-        facecolor="white",
-        edgecolor="none",
-    )
-
-
-def plot_write_sweep_formatted(ax: plt.Axes, dict_list: list[dict]):
-    plot_write_sweep(ax, dict_list)
-    ax.set_xlabel("$I_{\mathrm{write}}$ [$\mu$A]")
-    ax.set_ylabel("BER")
-    ax.set_xlim(0, 300)
-    return ax
-
-
-def plot_write_sweep_formatted_markers(ax: plt.Axes, data_dict: dict):
-    data = data_dict.get("data")
-    data2 = data_dict.get("data2")
-    colors = CMAP2(np.linspace(0, 1, 4))
-    ax.plot(
-        [d["write_current"] for d in data],
-        [d["write_temp"] for d in data],
-        "d",
-        color=colors[0],
-        markeredgecolor="black",
-        markeredgewidth=0.5,
-    )
-    ax.plot(
-        [d["write_current"] for d in data2],
-        [d["write_temp"] for d in data2],
-        "o",
-        color=colors[2],
-        markeredgecolor="black",
-        markeredgewidth=0.5,
-    )
-    ax.set_xlabel("$I_{\mathrm{write}}$ [$\mu$A]")
-    ax.set_ylabel("$T_{\mathrm{write}}$ [K]")
-    ax.set_xlim(0, 300)
-    ax.legend(
-        ["Lower bound", "Upper bound"],
-        loc="upper right",
-        fontsize=6,
-        facecolor="white",
-        frameon=True,
-    )
-    ax.grid()
-    return ax
-
-
-def plot_delay(ax: plt.Axes, data_dict: dict):
-    delay_list = data_dict.get("delay")
-    bit_error_rate = data_dict.get("bit_error_rate")
-    N = 200e3
-    sort_index = np.argsort(delay_list)
-    delay_list = np.array(delay_list)[sort_index]
-    bit_error_rate = np.array(bit_error_rate)[sort_index]
-    bit_error_rate = np.array(bit_error_rate).flatten()
-    ber_std = np.sqrt(bit_error_rate * (1 - bit_error_rate) / N)
-    ax.errorbar(
-        delay_list,
-        bit_error_rate,
-        yerr=ber_std,
-        fmt="-",
-        marker=".",
-        color="black",
-    )
-    ax.set_ylabel("BER")
-    ax.set_xlabel("Memory Retention Time (s)")
-
-    ax.set_xscale("log")
-    ax.set_xbound(lower=1e-6)
-    ax.grid(True, which="both", linestyle="--")
-
-    ax.set_yscale("log")
-    ax.set_ylim([1e-4, 1e-3])
-    ax.yaxis.set_minor_formatter(ticker.NullFormatter())
-
-
-def plot_ber_grid(ax: plt.Axes):
-    ARRAY_SIZE = (4, 4)
-    param_dict = initialize_dict(ARRAY_SIZE)
-    xloc_list = []
-    yloc_list = []
-    for c in CELLS:
-        xloc, yloc = convert_cell_to_coordinates(c)
-        param_dict = process_cell(CELLS[c], param_dict, xloc, yloc)
-        xloc_list.append(xloc)
-        yloc_list.append(yloc)
-
-    plot_parameter_array(
-        ax,
-        xloc_list,
-        yloc_list,
-        param_dict["bit_error_rate"],
-        log=True,
-        cmap=plt.get_cmap("Blues").reversed(),
-    )
-
-    ax.xaxis.set_label_position("bottom")
-    ax.xaxis.set_ticks_position("bottom")
-    ax.set_xlabel("Column")
-    ax.set_ylabel("Row")
-    cax = ax.inset_axes([1.10, 0, 0.1, 1])
-    cbar = fig.colorbar(
-        ax.get_children()[0], cax=cax, orientation="vertical", label="minimum BER"
-    )
-
-    return ax
 
 
 def import_write_sweep_formatted() -> list[dict]:
@@ -275,104 +87,6 @@ def import_write_sweep_formatted_markers(dict_list) -> list[dict]:
     }
     return data_dict
 
-
-def set_pres_style(dpi=600, font_size=14, grid_alpha=0.4):
-    """
-    Apply a presentation-optimized Matplotlib style.
-
-    Parameters:
-        dpi (int): Figure DPI (for saved files).
-        font_size (int): Base font size for axes and labels.
-        grid_alpha (float): Grid line transparency.
-    """
-    set_inter_font()
-    plt.rcParams.update({
-        "figure.dpi": dpi,
-        "font.family": "Inter",
-        "figure.figsize": (6, 4),
-        "axes.titlesize": font_size + 4,
-        "axes.labelsize": font_size + 2,
-        "xtick.labelsize": font_size,
-        "ytick.labelsize": font_size,
-        "legend.fontsize": font_size,
-        "font.size": font_size,
-        "axes.grid": True,
-        "grid.linestyle": "--",
-        "grid.linewidth": 0.7,
-        "grid.alpha": grid_alpha,
-        "axes.edgecolor": "#333333",
-        "axes.linewidth": 1.2,
-        "lines.linewidth": 2.0,
-        "savefig.bbox": "tight",
-        "savefig.pad_inches": 0.2,
-        "xtick.major.size": 5,
-        "ytick.major.size": 5,
-    })
-
-
-def set_inter_font():
-    if os.name == "nt":  # Windows
-        font_path = r"C:\Users\ICE\AppData\Local\Microsoft\Windows\Fonts\Inter-VariableFont_opsz,wght.ttf"
-    elif os.name == "posix":
-        font_path = "/home/omedeiro/Inter-VariableFont_opsz,wght.ttf"
-    else:
-        font_path = None
-
-    if font_path and os.path.exists(font_path):
-        fm.fontManager.addfont(font_path)
-        mpl.rcParams["font.family"] = "Inter"
-
-
-def set_plot_style() -> None:
-    set_inter_font()
-    golden_ratio = (1 + 5**0.5) / 2  # ≈1.618
-    width = 3.5  # Example width in inches (single-column for Nature)
-    height = width / golden_ratio
-    plt.rcParams.update(
-        {
-            "figure.figsize": [width, height],
-            "pdf.fonttype": 42,
-            "ps.fonttype": 42,
-            "font.family": "serif",
-            "mathtext.fontset": "cm",
-            'font.size': 9,
-            'axes.titlesize': 9,
-            'axes.labelsize': 9,
-            'xtick.labelsize': 8,
-            'ytick.labelsize': 8,
-            'legend.fontsize': 8,
-            "axes.linewidth": 0.5,
-            "xtick.major.width": 0.5,
-            "ytick.major.width": 0.5,
-            "xtick.direction": "out",
-            "ytick.direction": "out",
-            "lines.markersize": 3,
-            "lines.linewidth": 1.2,
-            "legend.frameon": False,
-            "xtick.major.size": 2,
-            "ytick.major.size": 2,
-        }
-    )
-
-
-set_plot_style()
-# COLORS = [
-#     "#1b9e77",  # Teal green
-#     "#d95f02",  # Burnt orange
-#     "#7570b3",  # Muted blue-purple
-#     "#e7298a",  # Reddish pink
-#     "#66a61e",  # Olive green
-#     "#e6ab02",  # Mustard yellow
-#     "#a6761d",  # Brown
-#     "#666666",  # Dark gray
-# ]
-
-# plt.rcParams["axes.prop_cycle"] = cycler(color=COLORS)
-C0 = "#1b9e77"
-C1 = "#d95f02"
-CMAP = plt.get_cmap("coolwarm")
-CMAP2 = mcolors.LinearSegmentedColormap.from_list("custom_cmap", [C0, C1])
-CMAP3 = plt.get_cmap("plasma").reversed()
 
 
 def build_array(
@@ -895,44 +609,6 @@ def get_state_currents_measured_array(
     return np.array(temps), np.array(state_currents)
 
 
-# def get_inverting_state_currents_measured(
-#     data_dict: dict,
-# ) -> Tuple[np.ndarray, np.ndarray]:
-#     bit_error_rate = get_bit_error_rate(data_dict)
-#     inverting_state_currents_list = []
-#     inverting_read_temperature_list = []
-#     _, _, inverting_edge1, inverting_edge2 = get_bit_error_rate_args(bit_error_rate)
-#     read_temperature = get_read_temperature(data_dict)
-#     read_currents = get_read_currents(data_dict)
-#     if inverting_edge1 is not np.nan:
-#         inverting_state0_current = read_currents[inverting_edge2]
-#         inverting_state1_current = read_currents[inverting_edge1]
-#         inverting_state_currents_list.append(
-#             [inverting_state1_current, inverting_state0_current]
-#         )
-#         inverting_read_temperature_list.append(read_temperature)
-#     else:
-#         inverting_state_currents_list.append([np.nan, np.nan])
-#         inverting_read_temperature_list.append(read_temperature)
-
-#     temp = np.array(inverting_state_currents_list)
-#     state_currents = np.array(inverting_state_currents_list)
-#     return temp, state_currents
-
-
-# def get_state_currents_measured(data_dict: dict) -> np.ndarray:
-#     nom_temp_array, nominal_state_current_array = get_nominal_state_currents_measured(
-#         data_dict
-#     )
-#     inv_temp_array, inverting_state_currents_array = (
-#         get_inverting_state_currents_measured(data_dict)
-#     )
-#     print(nom_temp_array)
-#     state_currents = np.array(
-#         [nominal_state_current_array, inverting_state_currents_array]
-#     )
-#     return state_currents
-
 
 def get_state_currents_array(dict_list: list[dict]) -> np.ndarray:
     state_currents = []
@@ -1036,1210 +712,6 @@ def process_cell(cell: dict, param_dict: dict, x: int, y: int) -> dict:
     return param_dict
 
 
-def plot_bit_error_rate_args(ax: Axes, data_dict: dict, color) -> Axes:
-    bit_error_rate = get_bit_error_rate(data_dict)
-    berargs = get_bit_error_rate_args(bit_error_rate)
-
-    read_current = get_read_currents(data_dict)
-    for arg in berargs:
-        if arg is not np.nan:
-            ax.plot(
-                read_current[arg],
-                bit_error_rate[arg],
-                marker="o",
-                color=color,
-            )
-            ax.axvline(read_current[arg], color=color, linestyle="--")
-    return ax
-
-
-def plot_bit_error_rate(ax: Axes, data_dict: dict) -> Axes:
-    bit_error_rate = get_bit_error_rate(data_dict)
-    total_switches_norm = get_total_switches_norm(data_dict)
-    measurement_param = data_dict.get("y")[0][:, 1] * 1e6
-
-    ax.plot(
-        measurement_param,
-        bit_error_rate,
-        color="#08519C",
-        label="Bit Error Rate",
-        marker=".",
-    )
-    ax.plot(
-        measurement_param,
-        total_switches_norm,
-        color="grey",
-        label="_Total Switches",
-        linestyle="--",
-        linewidth=1,
-    )
-    ax.legend()
-    ax.set_yticks([0, 0.5, 1])
-    ax.set_ylim([0, 1])
-    ax.set_ylabel("Normalized\nBit Error Rate")
-    return ax
-
-
-def plot_branch_currents(
-    ax: Axes,
-    T: np.ndarray,
-    Tc: float,
-    retrap_ratio: float,
-    width_ratio: float,
-    critical_current_zero: float,
-) -> Axes:
-    ichl, irhl, ichr, irhr = calculate_branch_currents(
-        T, Tc, retrap_ratio, width_ratio, critical_current_zero
-    )
-
-    ax.plot(T, ichl, label="$I_{c, H_L}(T)$", color="b", linestyle="-")
-    ax.plot(T, irhl, label="$I_{r, H_L}(T)$", color="b", linestyle="--")
-    ax.plot(T, ichr, label="$I_{c, H_R}(T)$", color="r", linestyle="-")
-    ax.plot(T, irhr, label="$I_{r, H_R}(T)$", color="r", linestyle="--")
-
-    ax.plot(T, ichr + irhl, label="$I_{0}(T)$", color="g", linestyle="-")
-    print(f"ichr: {ichr[0]}, irhl: {irhl[0]}")
-    print(f"sum {ichr[0]+irhl[0]}")
-    return ax
-
-
-def plot_channel_temperature(
-    ax: plt.Axes,
-    data_dict: dict,
-    channel_sweep: Literal["enable_write_current", "enable_read_current"],
-    **kwargs,
-) -> plt.Axes:
-    if channel_sweep == "enable_write_current":
-        temp = get_channel_temperature(data_dict, "write")
-        current = get_enable_current_sweep(data_dict, "write")
-        print(f"temp: {temp}, current: {current}")
-    else:
-        temp = get_channel_temperature(data_dict, "read")
-        current = get_enable_current_sweep(data_dict, "read")
-        print(f"2temp: {temp}, current: {current}")
-    ax.plot(current, temp, **kwargs)
-
-    return ax
-
-
-def plot_calculated_filled_region(
-    ax,
-    temp: np.ndarray,
-    data_dict: dict,
-    persistent_current: float,
-    critical_temp: float,
-    retrap_ratio: float,
-    width_ratio: float,
-    alpha: float,
-    critical_current_zero: float,
-) -> Axes:
-
-    plot_calculated_nominal_region(
-        ax,
-        temp,
-        data_dict,
-        persistent_current,
-        critical_temp,
-        retrap_ratio,
-        width_ratio,
-        alpha,
-        critical_current_zero,
-    )
-    plot_calculated_inverting_region(
-        ax,
-        temp,
-        data_dict,
-        persistent_current,
-        critical_temp,
-        retrap_ratio,
-        width_ratio,
-        alpha,
-        critical_current_zero,
-    )
-
-    return ax
-
-
-def plot_calculated_nominal_region(
-    ax: Axes,
-    temp: np.ndarray,
-    data_dict: dict,
-    persistent_current: float,
-    critical_temp: float,
-    retrap_ratio: float,
-    width_ratio: float,
-    alpha: float,
-    critical_current_zero: float,
-) -> Axes:
-    i0, i1, i2, i3 = calculate_state_currents(
-        temp,
-        critical_temp,
-        retrap_ratio,
-        width_ratio,
-        alpha,
-        persistent_current,
-        critical_current_zero,
-    )
-
-    upper_bound = np.minimum(i0, critical_current_zero)
-    lower_bound = np.maximum(np.minimum(np.maximum(i3, i1), critical_current_zero), i2)
-    ax.fill_between(
-        temp,
-        lower_bound,
-        upper_bound,
-        color="blue",
-        alpha=0.1,
-        hatch="////",
-    )
-    return ax
-
-
-def plot_calculated_inverting_region(
-    ax: Axes,
-    temp: np.ndarray,
-    data_dict: dict,
-    persistent_current: float,
-    critical_temp: float,
-    retrap_ratio: float,
-    width_ratio: float,
-    alpha: float,
-    critical_current_zero: float,
-) -> Axes:
-    i0, i1, i2, i3 = calculate_state_currents(
-        temp,
-        critical_temp,
-        retrap_ratio,
-        width_ratio,
-        alpha,
-        persistent_current,
-        critical_current_zero,
-    )
-    upper_bound = np.minimum(np.minimum(np.maximum(i0, i2), i1), critical_current_zero)
-    lower_bound = np.minimum(np.minimum(np.minimum(i2, i3), i0), critical_current_zero)
-    ax.fill_between(
-        temp,
-        lower_bound,
-        upper_bound,
-        color="red",
-        alpha=0.1,
-        hatch="\\\\\\\\",
-    )
-    return ax
-
-
-def plot_calculated_state_currents(
-    ax: Axes,
-    T: np.ndarray,
-    Tc: float,
-    retrap_ratio: float,
-    width_ratio: float,
-    alpha: float,
-    persistent_current: float,
-    critical_current_zero: float,
-    **kwargs,
-):
-    i0, i1, i2, i3 = calculate_state_currents(
-        T,
-        Tc,
-        retrap_ratio,
-        width_ratio,
-        alpha,
-        persistent_current,
-        critical_current_zero,
-    )
-    ax.plot(T, i0, label="$I_{{0}}(T)$", **kwargs)
-    ax.plot(T, i1, label="$I_{{1}}(T)$", **kwargs)
-    ax.plot(T, i2, label="$I_{{0,inv}}(T)$", **kwargs)
-    ax.plot(T, i3, label="$I_{{1,inv}}(T)$", **kwargs)
-    return ax
-
-
-def plot_cell_parameter(ax: Axes, param: str) -> Axes:
-    param_array = np.array([CELLS[cell][param] for cell in CELLS]).reshape(4, 4)
-    plot_parameter_array(
-        ax,
-        np.arange(4),
-        np.arange(4),
-        param_array * 1e6,
-        f"Cell {param}",
-        log=False,
-        norm=False,
-        reverse=False,
-    )
-    return ax
-
-
-def plot_critical_currents_from_trace(ax: Axes, dict_list: list) -> Axes:
-    critical_currents, critical_currents_std = get_critical_currents_from_trace(
-        dict_list
-    )
-    cmap = plt.cm.coolwarm(np.linspace(0, 1, len(dict_list)))
-    heater_currents = [data["heater_current"].flatten() * 1e6 for data in dict_list]
-    ax.errorbar(
-        heater_currents,
-        critical_currents,
-        yerr=critical_currents_std,
-        fmt="o",
-        markersize=3,
-        color=cmap[0, :],
-    )
-    ax.tick_params(direction="in", top=True, right=True)
-    ax.set_xlabel("Heater Current [µA]")
-    ax.set_ylabel("Critical Current [µA]")
-    ax.set_ylim([0, 400])
-    return ax
-
-
-def plot_critical_currents_from_dc_sweep(
-    ax: Axes, dict_list: list, substrate_temp: float = 1.3, save: bool = False
-) -> Axes:
-
-    critical_currents, critical_currents_std = get_critical_currents_from_trace(
-        dict_list
-    )
-
-    cmap = plt.cm.coolwarm(np.linspace(0, 1, len(dict_list)))
-    heater_currents = np.array(
-        [data["heater_current"].flatten() * 1e6 for data in dict_list]
-    ).flatten()
-    heater_currents = np.round(heater_currents, 0).astype(int)
-
-    ax.plot(
-        heater_currents,
-        critical_currents,
-        "o--",
-        color=cmap[0],
-        label="$I_{{EN}}$",
-        linewidth=0.5,
-        markersize=0.5,
-        markerfacecolor=cmap[0],
-    )
-
-    ax.fill_between(
-        heater_currents,
-        critical_currents + critical_currents_std,
-        critical_currents - critical_currents_std,
-        color=cmap[0],
-        alpha=0.3,
-        edgecolor="none",
-    )
-
-    ax.set_xlabel("$I_{{EN}}$ [µA]")
-    ax.set_ylabel("$I_{{c}}$ [µA]", labelpad=-1)
-    ax.set_ylim([0, 400])
-    ax.set_xlim([-500, 500])
-
-    # Add secondary y-axis for temperature
-    ax2 = ax.twinx()
-    temp = calculate_channel_temperature(
-        CRITICAL_TEMP, substrate_temp, np.abs(heater_currents), 500
-    )
-    ax2.plot(heater_currents, temp, "o--")
-    ax2.set_ylim([0, 13])
-    ax2.set_ylabel("Temperature [K]")
-    ax2.axhline(substrate_temp, color="black", linestyle="--", linewidth=0.5)
-    ax2.axhline(CRITICAL_TEMP, color="black", linestyle="--", linewidth=0.5)
-
-    if save:
-        plt.savefig("critical_currents_full.pdf", bbox_inches="tight")
-
-    return ax
-
-
-def plot_enable_current_relation(ax: Axes, dict_list: list[dict]) -> Axes:
-    colors = CMAP(np.linspace(0.1, 1, 4))
-    markers = ["o", "s", "D", "^"]
-    for data_dict in dict_list:
-        cell = get_current_cell(data_dict)
-        column, row = convert_cell_to_coordinates(cell)
-        x, y, ztotal = build_array(data_dict, "total_switches_norm")
-        xfit, yfit = get_fitting_points(x, y, ztotal)
-        ax.plot(xfit, yfit, label=f"{cell}", color=colors[column], marker=markers[row])
-    return ax
-
-
-def plot_fill_between(ax, data_dict, fill_color):
-    # fill the area between 0.5 and the curve
-    enable_write_currents = get_enable_current_sweep(data_dict)
-    bit_error_rate = get_bit_error_rate(data_dict)
-    verts = polygon_nominal(enable_write_currents, bit_error_rate)
-    poly = PolyCollection([verts], facecolors=fill_color, alpha=0.3, edgecolors="k")
-    ax.add_collection(poly)
-    verts = polygon_inverting(enable_write_currents, bit_error_rate)
-    poly = PolyCollection([verts], facecolors=fill_color, alpha=0.3, edgecolors="k")
-    ax.add_collection(poly)
-
-    return ax
-
-
-def plot_fitting(ax: Axes, xfit: np.ndarray, yfit: np.ndarray, **kwargs) -> Axes:
-    # xfit, yfit = filter_plateau(xfit, yfit, 0.98 * Ic0)
-    ax.plot(xfit, yfit, **kwargs)
-    plot_linear_fit(ax, xfit, yfit)
-
-    return ax
-
-
-def plot_linear_fit(
-    ax: Axes, xfit: np.ndarray, yfit: np.ndarray, add_text: bool = False
-) -> Axes:
-    z = np.polyfit(xfit, yfit, 1)
-    p = np.poly1d(z)
-    x_intercept = -z[1] / z[0]
-    # ax.scatter(xfit, yfit, color="#08519C")
-    xplot = np.linspace(0, x_intercept, 10)
-    ax.plot(xplot, p(xplot), ":", color="k")
-    if add_text:
-        ax.text(
-            0.1,
-            0.1,
-            f"{p[1]:.3f}x + {p[0]:.3f}\n$x_{{int}}$ = {x_intercept:.2f}",
-            fontsize=12,
-            color="red",
-            backgroundcolor="white",
-            transform=ax.transAxes,
-        )
-    ax.set_xlim((0, 600))
-    ax.set_ylim((0, 2000))
-
-    return ax
-
-
-def plot_measured_state_current_list(ax: Axes, dict_list: list[dict]) -> Axes:
-    sweep_length = len(dict_list)
-    for j in range(0, sweep_length):
-        plot_state_currents_measured(ax, dict_list[j], "enable_read_current")
-
-    return ax
-
-
-def plot_message(ax: Axes, message: str) -> Axes:
-    axheight = ax.get_ylim()[1]
-    for i, bit in enumerate(message):
-        text = get_text_from_bit_v3(bit)
-        ax.text(i + 0.5, axheight * 0.85, text, ha="center", va="center")
-
-    return ax
-
-
-def plot_parameter_array(
-    ax: Axes,
-    xloc: np.ndarray,
-    yloc: np.ndarray,
-    parameter_array: np.ndarray,
-    title: str = None,
-    log: bool = False,
-    reverse: bool = False,
-    cmap: plt.cm = None,
-) -> Axes:
-    if cmap is None:
-        cmap = plt.get_cmap("viridis")
-    if reverse:
-        cmap = cmap.reversed()
-
-    if log:
-        ax.matshow(
-            parameter_array,
-            cmap=cmap,
-            norm=LogNorm(vmin=np.min(parameter_array), vmax=np.max(parameter_array)),
-        )
-    else:
-        ax.matshow(parameter_array, cmap=cmap)
-
-    if title:
-        ax.set_title(title)
-    ax.set_xticks(range(4), ["A", "B", "C", "D"])
-    ax.set_yticks(range(4), ["1", "2", "3", "4"])
-    ax.tick_params(axis="both", length=0)
-    return ax
-
-
-def plot_enable_write_sweep_multiple(
-    ax: Axes,
-    dict_list: list[dict],
-    add_errorbar: bool = False,
-    N: int = None,
-    add_colorbar: bool = True,
-) -> Axes:
-    if N is None:
-        N = len(dict_list)
-    colors = CMAP(np.linspace(0, 1, N))
-    write_current_list = []
-    for data_dict in dict_list:
-        write_current = get_write_current(data_dict)
-        write_current_list.append(write_current)
-
-    for i, data_dict in enumerate(dict_list):
-        write_current_norm = write_current_list[i] / 100
-        plot_enable_sweep_single(
-            ax, data_dict, color=CMAP(write_current_norm), add_errorbar=add_errorbar
-        )
-
-    if add_colorbar:
-        norm = mcolors.Normalize(
-            vmin=min(write_current_list), vmax=max(write_current_list)
-        )  # Normalize for colormap
-        sm = plt.cm.ScalarMappable(cmap=CMAP, norm=norm)
-        sm.set_array([])
-        cbar = plt.colorbar(sm, ax=ax, orientation="vertical", fraction=0.05, pad=0.05)
-        cbar.set_label("Write Current [µA]")
-
-    ax.set_ylim(0, 1)
-    ax.yaxis.set_major_locator(MultipleLocator(0.5))
-    return ax
-
-
-def plot_enable_sweep_single(
-    ax: Axes,
-    data_dict: dict,
-    add_errorbar: bool = False,
-    **kwargs,
-) -> Axes:
-    enable_currents = get_enable_current_sweep(data_dict)
-    bit_error_rate = get_bit_error_rate(data_dict)
-    write_current = get_write_current(data_dict)
-    ax.plot(
-        enable_currents,
-        bit_error_rate,
-        label=f"$I_{{W}}$ = {write_current:.1f}µA",
-        **kwargs,
-    )
-    if add_errorbar:
-        # ax.errorbar(
-        #     enable_currents,
-        #     bit_error_rate,
-        #     yerr=np.sqrt(bit_error_rate * (1 - bit_error_rate) / len(bit_error_rate)),
-        #     **kwargs,
-        # )
-        ax.fill_between(
-            enable_currents,
-            bit_error_rate
-            - np.sqrt(bit_error_rate * (1 - bit_error_rate) / len(bit_error_rate)),
-            bit_error_rate
-            + np.sqrt(bit_error_rate * (1 - bit_error_rate) / len(bit_error_rate)),
-            alpha=0.1,
-            color=kwargs.get("color"),
-        )
-    ax.set_xlim(enable_currents[0], enable_currents[-1])
-    ax.set_ylim(0, 1)
-    ax.yaxis.set_major_locator(MultipleLocator(0.5))
-    ax.xaxis.set_major_locator(MultipleLocator(25))
-    ax.xaxis.set_minor_locator(MultipleLocator(5))
-    return ax
-
-def plot_ber_3d_bar(ber_array: np.ndarray, total_trials: int = 200_000) -> None:
-    barray = ber_array.copy()
-    fig = plt.figure(figsize=(7, 7))
-    ax = fig.add_subplot(111, projection="3d")
-
-    # 4x4 grid
-    x_data, y_data = np.meshgrid(np.arange(4), np.arange(4))
-    x_data = x_data.flatten()
-    y_data = y_data.flatten()
-    dz = barray.flatten()
-
-    # Mask invalid entries
-    valid_mask = np.isfinite(dz) & (dz < 5.5e-2)
-    x_data, y_data, dz = x_data[valid_mask], y_data[valid_mask], dz[valid_mask]
-
-    # Compute error counts
-    error_counts = np.round(dz * total_trials).astype(int)
-    z_data = np.zeros_like(error_counts)
-    dx = dy = 0.6 * np.ones_like(error_counts)
-    print(error_counts)
-    # Normalize error counts for colormap
-    norm = Normalize(vmin=error_counts.min(), vmax=error_counts.max())
-    colors = cm.Blues(norm(error_counts))
-
-    # Plot 3D bars with color
-    ax.bar3d(
-        x_data,
-        y_data,
-        z_data,
-        dx,
-        dy,
-        error_counts,
-        shade=True,
-        color=colors,
-        edgecolor="black",
-        linewidth=0.5,
-        alpha=0.8,
-    )
-
-    # Invert y-axis for logical row order
-    ax.invert_yaxis()
-
-    # Labels and ticks
-    ax.set_xlabel("Column")
-    ax.set_ylabel("Row")
-    ax.set_zlabel("Errors")
-    ax.set_title("Error Count per Cell")
-    ax.set_xticks(np.arange(4) + 0.3)  # Offset tick locations
-    ax.set_xticklabels(["A", "B", "C", "D"])
-    ax.set_yticks(np.arange(4) + 0.3)  # Offset tick locations
-    ax.set_yticklabels(["1", "2", "3", "4"])
-    ax.set_xlim(-0.5, 3.5)
-    ax.set_ylim(-0.5, 3.5)
-    ax.set_zlim(1, 500)
-    ax.view_init(elev=45, azim=220)
-
-    # Colorbar
-    mappable = cm.ScalarMappable(norm=norm, cmap=cm.Blues)
-    mappable.set_array([])
-    cbar = fig.colorbar(mappable, ax=ax, shrink=0.6, pad=0.1)
-    cbar.set_label("Errors (per 200k)")
-    fig.tight_layout()
-    fig.patch.set_visible(False)
-    save_fig = False
-    if save_fig:
-        fig.savefig("ber_3d_bar.pdf", dpi=300, bbox_inches="tight")
-    plt.show()
-
-
-# range set 1 [::2]
-def plot_enable_sweep(
-    ax: plt.Axes,
-    dict_list: list[dict],
-    range=None,
-    add_errorbar=False,
-    add_colorbar=False,
-):
-    N = len(dict_list)
-    if range is not None:
-        dict_list = dict_list[range]
-    # ax, ax2 = plot_enable_write_sweep_multiple(ax, dict_list[0:6])
-    ax = plot_enable_write_sweep_multiple(
-        ax, dict_list, add_errorbar=add_errorbar, N=N, add_colorbar=add_colorbar
-    )
-
-    ax.set_ylabel("BER")
-    ax.set_xlabel("$I_{\mathrm{enable}}$ [$\mu$A]")
-    return ax
-
-
-def plot_enable_sweep_markers(ax: plt.Axes, dict_list: list[dict]):
-    ax.yaxis.set_major_locator(plt.MultipleLocator(0.20))
-    ax.set_ylim([8.3, 9.7])
-
-    write_temp_array = np.empty((len(dict_list), 4))
-    write_current_array = np.empty((len(dict_list), 1))
-    enable_current_array = np.empty((len(dict_list), 4))
-    for j, data_dict in enumerate(dict_list):
-        bit_error_rate = get_bit_error_rate(data_dict)
-        berargs = get_bit_error_rate_args(bit_error_rate)
-        write_current = get_write_current(data_dict)
-        write_temps = get_channel_temperature_sweep(data_dict)
-        enable_currents = get_enable_current_sweep(data_dict)
-        write_current_array[j] = write_current
-        critical_current_zero = get_critical_current_heater_off(data_dict)
-        for i, arg in enumerate(berargs):
-            if arg is not np.nan:
-                write_temp_array[j, i] = write_temps[arg]
-                enable_current_array[j, i] = enable_currents[arg]
-    markers = ["o", "s", "D", "^"]
-    for i in range(4):
-        ax.plot(
-            enable_current_array[:, i],
-            write_current_array,
-            linestyle="--",
-            marker=markers[i],
-            markeredgecolor="k",
-            markeredgewidth=0.5,
-            color=RBCOLORS[i],
-        )
-    ax.set_ylim(0, 100)
-    ax.set_xlim(250, 340)
-    ax.yaxis.set_major_locator(plt.MultipleLocator(20))
-    ax.yaxis.set_minor_locator(plt.MultipleLocator(10))
-    ax.xaxis.set_major_locator(plt.MultipleLocator(25))
-    ax.xaxis.set_minor_locator(plt.MultipleLocator(5))
-    ax.grid()
-    ax.set_ylabel("$I_{\mathrm{write}}$ [$\mu$A]")
-    ax.set_xlabel("$I_{\mathrm{enable}}$ [$\mu$A]")
-    ax.legend(
-        [
-            "$I_{1}$",
-            "$I_{0}$",
-            "$I_{0,\mathrm{inv}}$",
-            "$I_{1,\mathrm{inv}}$",
-        ],
-        loc="lower left",
-        frameon=True,
-        ncol=1,
-        facecolor="white",
-        edgecolor="none",
-    )
-
-
-def plot_write_sweep_formatted(ax: plt.Axes, dict_list: list[dict]):
-    plot_write_sweep(ax, dict_list)
-    ax.set_xlabel("$I_{\mathrm{write}}$ [$\mu$A]")
-    ax.set_ylabel("BER")
-    ax.set_xlim(0, 300)
-    return ax
-
-
-def plot_write_sweep_formatted_markers(ax: plt.Axes, data_dict: dict):
-    data = data_dict.get("data")
-    data2 = data_dict.get("data2")
-    colors = CMAP2(np.linspace(0, 1, 4))
-    ax.plot(
-        [d["write_current"] for d in data],
-        [d["write_temp"] for d in data],
-        "d",
-        color=colors[0],
-        markeredgecolor="black",
-        markeredgewidth=0.5,
-    )
-    ax.plot(
-        [d["write_current"] for d in data2],
-        [d["write_temp"] for d in data2],
-        "o",
-        color=colors[2],
-        markeredgecolor="black",
-        markeredgewidth=0.5,
-    )
-    ax.set_xlabel("$I_{\mathrm{write}}$ [$\mu$A]")
-    ax.set_ylabel("$T_{\mathrm{write}}$ [K]")
-    ax.set_xlim(0, 300)
-    ax.legend(
-        ["Lower bound", "Upper bound"],
-        loc="upper right",
-        fontsize=6,
-        facecolor="white",
-        frameon=True,
-    )
-    ax.grid()
-    return ax
-
-
-
-def plot_delay(ax: plt.Axes, data_dict: dict):
-    delay_list = np.array(data_dict.get("delay"))
-    bit_error_rate = np.array(data_dict.get("bit_error_rate")).flatten()
-    N = 200e3  # Total trials for BER standard deviation
-
-    # Sort data for proper line plotting
-    sort_index = np.argsort(delay_list)
-    delay_list = delay_list[sort_index]
-    bit_error_rate = bit_error_rate[sort_index]
-
-    # Error bars calculation
-    ber_std = np.sqrt(bit_error_rate * (1 - bit_error_rate) / N)
-
-    # Plotting with improved style
-    ax.errorbar(
-        delay_list,
-        bit_error_rate,
-        yerr=ber_std,
-        fmt="-o",
-        color="black",
-        elinewidth=1,
-        capsize=2,
-    )
-
-    # Axes labels with specific font sizes
-    ax.set_xlabel("$\Delta t$ [s]")
-    ax.set_ylabel("BER")
-
-    # Log scales with tick formatting
-    ax.set_xscale("log")
-    ax.set_yscale("log")
-    ax.set_xlim([1e-6, delay_list.max() * 1.1])
-    ax.set_ylim([1e-4, 1e-3])
-
-    # Major and minor ticks on log scale
-    ax.xaxis.set_major_locator(LogLocator(base=10.0, numticks=5))
-    ax.xaxis.set_minor_locator(
-        LogLocator(base=10.0, subs=np.arange(2, 10) * 0.1, numticks=10)
-    )
-    ax.xaxis.set_minor_formatter(NullFormatter())
-
-    ax.yaxis.set_major_locator(LogLocator(base=10.0, numticks=5))
-    ax.yaxis.set_minor_locator(
-        LogLocator(base=10.0, subs=np.arange(2, 10) * 0.1, numticks=10)
-    )
-    ax.yaxis.set_minor_formatter(NullFormatter())
-
-    # Gridlines
-    ax.grid(which="major", linestyle="--", linewidth=0.5, alpha=0.7)
-    ax.grid(which="minor", linestyle=":", linewidth=0.4, alpha=0.5)
-
-    # Tick parameters
-    ax.tick_params(axis="both", which="major", labelsize=9)
-    ax.tick_params(axis="both", which="minor", labelsize=7)
-
-
-def plot_ber_grid(ax: plt.Axes):
-    set_plot_style()
-    ARRAY_SIZE = (4, 4)
-    param_dict = initialize_dict(ARRAY_SIZE)
-    xloc_list = []
-    yloc_list = []
-    for c in CELLS:
-        xloc, yloc = convert_cell_to_coordinates(c)
-        param_dict = process_cell(CELLS[c], param_dict, xloc, yloc)
-        xloc_list.append(xloc)
-        yloc_list.append(yloc)
-
-    plot_parameter_array(
-        ax,
-        xloc_list,
-        yloc_list,
-        param_dict["bit_error_rate"],
-        log=True,
-        cmap=plt.get_cmap("Blues").reversed(),
-    )
-
-    ax.xaxis.set_label_position("bottom")
-    ax.xaxis.set_ticks_position("bottom")
-    ax.set_xlabel("Column")
-    ax.set_ylabel("Row")
-    cax = ax.inset_axes([1.10, 0, 0.1, 1])
-    fig = plt.gcf()
-    cbar = fig.colorbar(
-        ax.get_children()[0], cax=cax, orientation="vertical", label="minimum BER"
-    )
-
-    return ax
-
-
-def import_write_sweep_formatted() -> list[dict]:
-    dict_list = import_directory(
-        os.path.join(os.path.dirname(__file__), "write_current_sweep_enable_write/data")
-    )
-    dict_list = dict_list[1:]
-    dict_list = dict_list[::-1]
-    dict_list = sorted(
-        dict_list, key=lambda x: x.get("enable_write_current").flatten()[0]
-    )
-    return dict_list
-
-
-def import_delay_dict() -> dict:
-    dict_list = import_directory(
-        os.path.join(os.path.dirname(__file__), "read_delay_v2/data3")
-    )
-    delay_list = []
-    bit_error_rate_list = []
-    for data_dict in dict_list:
-        delay = data_dict.get("delay").flatten()[0] * 1e-3
-        bit_error_rate = get_bit_error_rate(data_dict)
-
-        delay_list.append(delay)
-        bit_error_rate_list.append(bit_error_rate)
-
-    delay_dict = {}
-    delay_dict["delay"] = delay_list
-    delay_dict["bit_error_rate"] = bit_error_rate_list
-    return delay_dict
-
-
-def import_write_sweep_formatted_markers(dict_list) -> list[dict]:
-    data = []
-    data2 = []
-    for data_dict in dict_list:
-        bit_error_rate = get_bit_error_rate(data_dict)
-        berargs = get_bit_error_rate_args(bit_error_rate)
-        write_currents = get_read_currents(
-            data_dict
-        )  # This is correct. "y" is the write current in this .mat.
-        enable_write_current = get_enable_write_current(data_dict)
-        read_current = get_read_current(data_dict)
-        write_current = get_write_current(data_dict)
-
-        for i, arg in enumerate(berargs):
-            if arg is not np.nan:
-
-                if i == 0:
-                    data.append(
-                        {
-                            "write_current": write_currents[arg],
-                            "write_temp": get_channel_temperature(data_dict, "write"),
-                            "read_current": read_current,
-                            "enable_write_current": enable_write_current,
-                        }
-                    )
-                if i == 2:
-                    data2.append(
-                        {
-                            "write_current": write_currents[arg],
-                            "write_temp": get_channel_temperature(data_dict, "write"),
-                            "read_current": read_current,
-                            "enable_write_current": enable_write_current,
-                        }
-                    )
-    data_dict = {
-        "data": data,
-        "data2": data2,
-    }
-    return data_dict
-
-
-def plot_fidelity_clean_bar(ber_array: np.ndarray, total_trials: int = 200_000) -> None:
-    import matplotlib.pyplot as plt
-    import numpy as np
-
-    # Prepare data
-    barray = ber_array.copy().T
-    ber_flat = barray.flatten()
-    valid_mask = np.isfinite(ber_flat)
-    ber_flat = ber_flat[valid_mask]
-    fidelity_flat = 1 - ber_flat
-
-    # Generate A1 to D4 labels
-    rows = ["A", "B", "C", "D"]
-    cols = range(1, 5)
-    all_labels = [f"{r}{c}" for r in rows for c in cols]
-    labels = np.array(all_labels)[valid_mask]
-
-    # Error bars from binomial standard deviation
-    errors = np.sqrt(ber_flat * (1 - ber_flat) / total_trials)
-    # Clean label values
-    display_values = [f"{f:.5f}" if f < 0.99999 else "≥0.99999" for f in fidelity_flat]
-
-    # Plotting
-    fig, ax = plt.subplots(figsize=(10, 3))
-    x = np.arange(len(fidelity_flat))
-    ax.bar(x, fidelity_flat, yerr=errors, capsize=3, color="#658DDC", edgecolor="black")
-
-    # Add value labels only if they fit in the visible range
-    for i, (val, label) in enumerate(zip(fidelity_flat, display_values)):
-        if val > 0.998:  # only plot label if within axis range
-            ax.text(i, val + 1e-4, label, ha="center", va="bottom", fontsize=8)
-        if val < 0.998:
-            ax.text(i, 0.998 + 2e-4, label, ha="center", va="top", fontsize=8)
-    # Formatting
-    ax.set_xticks(x)
-    ax.set_xlim(-1.5, len(x) + 0.5)
-    ax.set_xticklabels(labels, rotation=45, ha="right")
-    ax.set_ylabel("Fidelity (1 - BER)")
-    ax.set_ylim(0.998, 1.0001)
-    ax.grid(axis="y", linestyle="--", alpha=0.6)
-
-    # Reference lines
-    ax.axhline(0.999, color="#555555", linestyle="--", linewidth=0.8, zorder=3)
-    ax.axhline(0.9999, color="#555555", linestyle="--", linewidth=0.8, zorder=3)
-    ax.axhline(
-        1-1.5e-3,
-        color="red",
-        linestyle="--",
-        linewidth=0.8,
-        zorder=3,
-    )
-    ax.text(
-        len(x) + 0.3,
-        1-1.5e-3,
-        "Previous Record",
-        va="top",
-        ha="right",
-        fontsize=14,
-        color="red",
-        zorder=4,
-        bbox=dict(facecolor="white", edgecolor="none", alpha=0.8, pad=1),
-    )
-    ax.set_ylim(0.998, 1.0004)
-    ax.set_yticks([0.998, 0.999, 0.9999])
-    ax.set_yticklabels(["0.998", "0.999", "0.9999"])
-    fig.patch.set_visible(False)
-    save_fig = False
-    if save_fig:
-        plt.savefig(
-            "fidelity_clean_bar.pdf",
-            dpi=300,
-            bbox_inches="tight",
-        )
-        plt.tight_layout()
-    plt.show()
-
-
-
-def plot_read_sweep(
-    ax: Axes,
-    data_dict: dict,
-    value_name: Literal["bit_error_rate", "write_0_read_1", "write_1_read_0"],
-    variable_name: Literal[
-        "enable_write_current",
-        "read_width",
-        "write_width",
-        "write_current",
-        "enable_read_current",
-        "enable_write_width",
-    ],
-    add_errorbar: bool = False,
-    **kwargs,
-) -> Axes:
-    write_temp = None
-    label = None
-    read_currents = get_read_currents(data_dict)
-
-    if value_name == "bit_error_rate":
-        value = get_bit_error_rate(data_dict)
-    if value_name == "write_0_read_1":
-        value = data_dict.get("write_0_read_1").flatten()
-    if value_name == "write_1_read_0":
-        value = data_dict.get("write_1_read_0").flatten()
-
-    if variable_name == "write_current":
-        variable = get_write_current(data_dict)
-        label = f"{variable:.2f}µA"
-    if variable_name == "enable_write_current":
-        variable = get_enable_write_current(data_dict)
-        write_temp = get_channel_temperature(data_dict, "write")
-        if write_temp is None:
-            label = f"{variable:.2f}µA"
-        else:
-            label = f"{variable:.2f}µA, {write_temp:.2f}K"
-    if variable_name == "read_width":
-        variable = get_read_width(data_dict)
-        label = f"{variable:.2f} pts "
-    if variable_name == "write_width":
-        variable = get_write_width(data_dict)
-        label = f"{variable:.2f} pts "
-    if variable_name == "enable_read_current":
-        variable = get_enable_read_current(data_dict)
-        read_temp = get_channel_temperature(data_dict, "read")
-        label = f"{variable:.2f}µA, {read_temp:.2f}K"
-    if variable_name == "enable_write_width":
-        variable = get_enable_write_width(data_dict)
-        label = f"{variable:.2f} pts"
-
-    ax.plot(
-        read_currents,
-        value,
-        label=label,
-        **kwargs,
-    )
-    if add_errorbar:
-        # ax.errorbar(
-        #     read_currents,
-        #     value,
-        #     yerr=np.sqrt(value * (1 - value) / len(value)),
-        #     **kwargs,
-        # )
-        ax.fill_between(
-            read_currents,
-            value - np.sqrt(value * (1 - value) / len(value)),
-            value + np.sqrt(value * (1 - value) / len(value)),
-            alpha=0.1,
-            color=kwargs.get("color"),
-        )
-    ax.set_ylim(0, 1)
-    ax.yaxis.set_major_locator(MultipleLocator(0.5))
-    return ax
-
-
-def plot_read_sweep_switch_probability(
-    ax: Axes,
-    data_dict: dict,
-    **kwargs,
-) -> Axes:
-    read_currents = get_read_currents(data_dict)
-    _, _, total_switch_probability = build_array(data_dict, "total_switches_norm")
-    ax.plot(
-        read_currents,
-        total_switch_probability,
-        **kwargs,
-    )
-    ax.yaxis.set_major_locator(MultipleLocator(0.5))
-    ax.set_ylim(0, 1)
-    return ax
-
-
-def plot_fill_between_array(ax: Axes, dict_list: list[dict]) -> Axes:
-    colors = CMAP(np.linspace(0.1, 1, len(dict_list)))
-    for i, data_dict in enumerate(dict_list):
-        plot_fill_between(ax, data_dict, colors[i])
-    return ax
-
-
-def plot_read_sweep_array(
-    ax: Axes,
-    dict_list: list[dict],
-    value_name: str,
-    variable_name: str,
-    colorbar=None,
-    add_errorbar=False,
-    **kwargs,
-) -> Axes:
-    colors = CMAP(np.linspace(0, 1, len(dict_list)))
-    variable_list = []
-    for i, data_dict in enumerate(dict_list):
-        plot_read_sweep(
-            ax,
-            data_dict,
-            value_name,
-            variable_name,
-            color=colors[i],
-            add_errorbar=add_errorbar,
-            **kwargs,
-        )
-        variable_list.append(data_dict[variable_name].flatten()[0] * 1e6)
-
-    if colorbar is not None:
-        norm = mcolors.Normalize(
-            vmin=min(variable_list), vmax=max(variable_list)
-        )  # Normalize for colormap
-        sm = plt.cm.ScalarMappable(cmap=CMAP, norm=norm)
-        sm.set_array([])
-        cbar = plt.colorbar(sm, ax=ax, orientation="vertical", fraction=0.05, pad=0.05)
-        cbar.set_label("Write Current [µA]")
-
-    return ax
-
-
-def plot_read_switch_probability_array(
-    ax: Axes, dict_list: list[dict], write_list=None, **kwargs
-) -> Axes:
-    colors = CMAP(np.linspace(0, 1, len(dict_list)))
-    print(f"len dict_list: {len(dict_list)}")
-    for i, data_dict in enumerate(dict_list):
-        if write_list is not None:
-            plot_read_sweep_switch_probability(
-                ax,
-                data_dict,
-                color=colors[i],
-                label=f"{write_list[i]} µA",
-                **kwargs,
-            )
-        else:
-            plot_read_sweep_switch_probability(ax, data_dict, color=colors[i])
-    return ax
-
-
-def plot_read_delay(ax: Axes, dict_list: dict) -> Axes:
-    colors = CMAP(np.linspace(0.1, 1, len(dict_list)))
-    for i, data_dict in enumerate(dict_list):
-        read_currents = get_read_currents(data_dict)
-        bit_error_rate = get_bit_error_rate(data_dict)
-        ax.plot(
-            read_currents,
-            bit_error_rate,
-            label=f"+{i+1}µs",
-            color=colors[i],
-            marker=".",
-            markeredgecolor="k",
-        )
-    ax.set_xlim(read_currents[0], read_currents[-1])
-    ax.set_yscale("log")
-    return ax
-
-
-def plot_write_sweep(ax: Axes, dict_list: str) -> Axes:
-    # colors = CMAP(np.linspace(0.1, 1, len(dict_list)))
-    colors = CMAP(np.linspace(0.1, 1, len(dict_list)))
-    write_temp_list = []
-    enable_write_current_list = []
-    for i, data_dict in enumerate(dict_list):
-        x, y, ztotal = build_array(data_dict, "bit_error_rate")
-        _, _, zswitch = build_array(data_dict, "total_switches_norm")
-        write_temp = get_channel_temperature(data_dict, "write")
-        enable_write_current = get_enable_write_current(data_dict)
-        write_temp_list.append(write_temp)
-        enable_write_current_list.append(enable_write_current)
-        ax.plot(
-            y,
-            ztotal,
-            label=f"$T_{{W}}$ = {write_temp:.2f} K, $I_{{EW}}$ = {enable_write_current:.2f} µA",
-            color=colors[dict_list.index(data_dict)],
-            marker=".",
-        )
-    ax.set_ylim([0, 1])
-    ax.yaxis.set_major_locator(MultipleLocator(0.5))
-    norm = mcolors.Normalize(
-        vmin=min(enable_write_current_list), vmax=max(enable_write_current_list)
-    )
-    sm = plt.cm.ScalarMappable(cmap=CMAP, norm=norm)
-    sm.set_array([])
-    cbar = plt.colorbar(sm, ax=ax, orientation="vertical", fraction=0.05, pad=0.05)
-    cbar.set_label("$I_{{EW}}$ [µA]")
-    return ax
-
-
-def plot_threshold(ax: Axes, start: int, end: int, threshold: float) -> Axes:
-    ax.hlines(threshold, start, end, color="red", ls="-", lw=1)
-    return ax
-
-
-def plot_text_labels(
-    ax: Axes, xloc: np.ndarray, yloc: np.ndarray, ztotal: np.ndarray, log: bool
-) -> Axes:
-    for x, y in zip(xloc, yloc):
-        text = f"{ztotal[y, x]:.2f}"
-        txt_color = "black"
-        if ztotal[y, x] > (0.8 * max(ztotal.flatten())):
-            txt_color = "white"
-        if log:
-            text = f"{ztotal[y, x]:.1e}"
-            txt_color = "black"
-
-        ax.text(
-            x,
-            y,
-            text,
-            color=txt_color,
-            backgroundcolor="none",
-            ha="center",
-            va="center",
-            weight="bold",
-        )
-
-    return ax
-
-
-def plot_state_currents_measured_nominal(
-    ax: Axes, nominal_read_temperature_list: list, nominal_state_currents_list: list
-) -> Axes:
-    for t, temp in enumerate(nominal_read_temperature_list):
-        ax.plot(
-            [temp, temp],
-            nominal_state_currents_list[t],
-            "o",
-            linestyle="-",
-            color="blue",
-        )
-    return ax
-
-
-def plot_state_currents_measured_inverting(
-    ax: Axes, inverting_read_temperature_list: list, inverting_state_currents_list: list
-) -> Axes:
-    for t, temp in enumerate(inverting_read_temperature_list):
-        ax.plot(
-            [temp, temp],
-            inverting_state_currents_list[t],
-            "o",
-            linestyle="-",
-            color="red",
-        )
-    return ax
-
-
-def plot_state_currents_measured(ax: Axes, data_dict: dict, current_sweep: str) -> Axes:
-    temp, state_currents = get_state_currents_measured(data_dict, current_sweep)
-
-    if state_currents[0] is not np.nan:
-        ax.plot(
-            [temp, temp],
-            state_currents[0:2],
-            "o",
-            linestyle="-",
-            color="blue",
-            label="_state0",
-        )
-    if state_currents[2] is not np.nan:
-        ax.plot(
-            [temp, temp],
-            state_currents[2:4],
-            "o",
-            linestyle="-",
-            color="red",
-            label="_state1",
-        )
-
-    return ax
-
-
 def get_state_current_markers_list(
     dict_list: list[dict],
     current_sweep: Literal["read_current", "enable_write_current"],
@@ -2271,404 +743,6 @@ def get_state_current_markers(
 
     return state_current_markers
 
-
-def plot_state_current_markers(
-    ax: Axes,
-    data_dict: dict,
-    current_sweep: Literal["read_current", "enable_write_current"],
-    **kwargs,
-) -> Axes:
-
-    state_current_markers = get_state_current_markers(data_dict, current_sweep)
-    currents = state_current_markers[0, :]
-    bit_error_rate = state_current_markers[1, :]
-    if currents[0] > 0:
-        for i in range(2):
-            ax.plot(
-                currents[i],
-                bit_error_rate[i],
-                color="blue",
-                marker="o",
-                markeredgecolor="k",
-                linewidth=1.5,
-                label="_state0",
-                markersize=12,
-                **kwargs,
-            )
-    if currents[2] > 0:
-        for i in range(2, 4):
-            ax.plot(
-                currents[i],
-                bit_error_rate[i],
-                color="red",
-                marker="o",
-                markeredgecolor="k",
-                linewidth=1.5,
-                label="_state1",
-                markersize=12,
-                **kwargs,
-            )
-    return ax
-
-
-def plot_voltage_trace(
-    ax: Axes, time: np.ndarray, voltage: np.ndarray, **kwargs
-) -> Axes:
-    ax.plot(time, voltage, **kwargs)
-    # ax.set_xlim(time[0], time[-1])
-    ax.yaxis.set_major_locator(MaxNLocator(4))
-    ax.set_xticklabels([])
-    ax.xaxis.set_major_locator(MultipleLocator(5))
-    ax.xaxis.set_minor_locator(MultipleLocator(1))
-    ax.tick_params(axis="x", direction="in", which="both")
-    # ax.grid(axis="x", which="both")
-    return ax
-
-
-def plot_voltage_trace_zoom(
-    ax: Axes, x: np.ndarray, y: np.ndarray, start: float, end: float
-) -> Axes:
-    xzoom = x[(x > start) & (x < end)]
-    yzoom = y[(x > start) & (x < end)]
-
-    # smooth the yzoom data
-    yzoom = np.convolve(yzoom, np.ones(20) / 20, mode="same")
-    ax.plot(xzoom, 400 + yzoom * 10, color="red", ls="--", lw=1)
-    ax.hlines(400, start, end, color="grey", ls="--", lw=1)
-
-    return ax
-
-
-def plot_voltage_trace_stack(
-    axs: List[Axes], data_dict: dict, trace_index: int = 0
-) -> List[Axes]:
-    colors = CMAP(np.linspace(0.1, 1, 3))
-    colors = np.flipud(colors)
-    if len(axs) != 3:
-        raise ValueError("The number of axes must be 3.")
-
-    chan_in_x, chan_in_y = get_voltage_trace_data(
-        data_dict, "trace_chan_in", trace_index
-    )
-    chan_out_x, chan_out_y = get_voltage_trace_data(
-        data_dict, "trace_chan_out", trace_index
-    )
-    enab_in_x, enab_in_y = get_voltage_trace_data(data_dict, "trace_enab", trace_index)
-
-    bitmsg_channel = data_dict.get("bitmsg_channel")[0]
-    bitmsg_enable = data_dict.get("bitmsg_enable")[0]
-
-    plot_voltage_trace(axs[0], chan_in_x, chan_in_y, color=colors[0], label="Input")
-
-    if bitmsg_enable[1] == "W" and bitmsg_channel[1] != "N":
-        plot_voltage_trace_zoom(axs[0], chan_in_x, chan_in_y, 0.9, 2.1)
-        plot_voltage_trace_zoom(axs[0], chan_in_x, chan_in_y, 4.9, 6.1)
-
-    if bitmsg_enable[3] == "W" and bitmsg_channel[3] != "N":
-        plot_voltage_trace_zoom(axs[0], chan_in_x, chan_in_y, 2.9, 4.1)
-        plot_voltage_trace_zoom(axs[0], chan_in_x, chan_in_y, 6.9, 8.1)
-
-    plot_voltage_trace(axs[1], enab_in_x, enab_in_y, color=colors[1], label="Enable")
-
-    plot_voltage_trace(axs[2], chan_out_x, chan_out_y, color=colors[2], label="Output")
-
-    axs[2].xaxis.set_major_locator(MultipleLocator(5))
-    axs[2].xaxis.set_major_formatter(plt.FuncFormatter(lambda x, _: f"{int(x)}"))
-    axs[2].set_xlim([0, 10])
-
-    fig = plt.gcf()
-    fig.supylabel("Voltage [mV]")
-    fig.supxlabel("Time [µs]")
-    fig.subplots_adjust(hspace=0.0)
-
-    return axs
-
-
-def plot_voltage_trace_averaged(
-    ax: Axes, data_dict: dict, trace_name: str, **kwargs
-) -> Axes:
-    x, y = get_voltage_trace_data(data_dict, trace_name)
-    ax.plot(
-        (x - x[0]),
-        y,
-        **kwargs,
-    )
-    return ax
-
-
-def plot_voltage_hist(ax: Axes, data_dict: dict) -> Axes:
-    ax.hist(
-        data_dict["read_zero_top"][0, :]*1e3,
-        log=True,
-        range=(200,600),
-        bins=100,
-        label="Read 0",
-        color="#658DDC",
-        alpha=0.8,
-        zorder=-1,
-    )
-    ax.hist(
-        data_dict["read_one_top"][0, :]*1e3,
-        log=True,
-        range=(200, 600),
-        bins=100,
-        label="Read 1",
-        color="#DF7E79",
-        alpha=0.8,
-    )
-    # ax.set_xlabel("$V$ [mV]")
-    # ax.set_ylabel("$N$")
-    ax.legend()
-    return ax
-
-
-def plot_voltage_trace_bitstream(ax: Axes, data_dict: dict, trace_name: str) -> Axes:
-    x, y = get_voltage_trace_data(
-        data_dict,
-        trace_name,
-    )
-    ax.plot(
-        x,
-        y,
-        label=trace_name,
-    )
-    plot_message(ax, data_dict["bitmsg_channel"][0])
-    return ax
-
-
-def plot_current_voltage_curve(ax: Axes, data_dict: dict, **kwargs) -> Axes:
-    time = data_dict.get("trace")[0, :]
-    voltage = data_dict.get("trace")[1, :]
-
-    M = int(np.round(len(voltage), -2))
-    currentQuart = np.linspace(0, data_dict["vpp"] / 2 / 10e3, M // 4)
-    current = np.concatenate(
-        [-currentQuart, np.flip(-currentQuart), currentQuart, np.flip(currentQuart)]
-    )
-
-    if len(voltage) > M:
-        voltage = voltage[:M]
-        time = time[:M]
-    else:
-        voltage = np.concatenate([voltage, np.zeros(M - len(voltage))])
-        time = np.concatenate([time, np.zeros(M - len(time))])
-
-    ax.plot(voltage, current.flatten() * 1e6, **kwargs)
-    return ax
-
-
-def plot_current_voltage_from_dc_sweep(
-    ax: Axes, dict_list: list, save: bool = False
-) -> Axes:
-    colors = plt.cm.coolwarm(np.linspace(0, 1, int(len(dict_list) / 2) + 1))
-    colors = np.flipud(colors)
-    for i, data in enumerate(dict_list):
-        heater_current = np.abs(data["heater_current"].flatten()[0] * 1e6)
-        ax = plot_current_voltage_curve(
-            ax, data, color=colors[i], zorder=-i, label=f"{heater_current:.0f} µA"
-        )
-        if i == 10:
-            break
-    ax.set_ylim([-500, 500])
-    ax.set_xlabel("Voltage [V]")
-    ax.set_ylabel("Current [µA]")
-
-    if save:
-        plt.savefig("iv_curve.pdf", bbox_inches="tight")
-
-    return ax
-
-
-def plot_optimal_enable_currents(ax: Axes, data_dict: dict) -> Axes:
-    cell = get_current_cell(data_dict)
-    enable_read_current = get_optimal_enable_read_current(cell)
-    enable_write_current = get_optimal_enable_write_current(cell)
-    ax.vlines(
-        [enable_write_current],
-        *ax.get_ylim(),
-        linestyle="--",
-        color="grey",
-        label="_Enable Write Current",
-    )
-    ax.vlines(
-        [enable_read_current],
-        *ax.get_ylim(),
-        linestyle="--",
-        color="r",
-        label="_Enable Read Current",
-    )
-    return ax
-
-
-def plot_grid(axs: Axes, dict_list: list[dict]) -> Axes:
-    colors = CMAP3(np.linspace(0.1, 1, 4))
-    markers = ["o", "s", "D", "^"]
-    for data_dict in dict_list:
-        cell = get_current_cell(data_dict)
-
-        column, row = convert_cell_to_coordinates(cell)
-        x = data_dict["x"][0]
-        y = data_dict["y"][0]
-        ztotal = data_dict["ztotal"]
-        xfit, yfit = get_fitting_points(x, y, ztotal)
-        axs[row, column].plot(
-            xfit,
-            yfit,
-            label=f"{cell}",
-            color=colors[column],
-            marker=markers[row],
-        )
-        dd = sio.loadmat(
-            "/home/omedeiro/nmem/src/nmem/analysis/enable_current_relation_v2/data2/SPG806_20250107_nMem_measure_enable_response_D6_A4_A3_2025-01-07 15-34-05.mat"
-        )
-        y_step_size = y[1] - y[0]
-        # print(x)
-        # print(y)
-        # axs[row, column].errorbar(
-        #     xfit,
-        #     yfit,
-        #     yerr=y_step_size * np.ones_like(yfit),
-        #     fmt="o",
-        #     color=colors[column],
-        #     marker=markers[row],
-        #     markeredgecolor="k",
-        #     markeredgewidth=0.1,
-        #     markersize=5,
-        #     alpha=0.5,
-        #     zorder=1,
-        #     label="_data",
-        # )
-
-        xfit, yfit = filter_plateau(xfit, yfit, yfit[0] * 0.75)
-
-        plot_linear_fit(
-            axs[row, column],
-            xfit,
-            yfit,
-        )
-        # plot_optimal_enable_currents(axs[row, column], data_dict)
-        axs[row, column].legend(loc="upper right")
-        axs[row, column].xaxis.set_major_locator(MultipleLocator(500))
-        axs[row, column].xaxis.set_minor_locator(MultipleLocator(100))
-        axs[row, column].set_xlim(0, 600)
-        axs[row, column].set_ylim(0, 1500)
-    axs[-1, 0].set_xlabel("Enable Current [µA]")
-    axs[-1, 0].set_ylabel("Critical Current [µA]")
-    return axs
-
-
-def plot_row(axs, dict_list):
-    colors = CMAP3(np.linspace(0.1, 1, 4))
-    markers = ["o", "s", "D", "^"]
-    for data_dict in dict_list:
-        cell = get_current_cell(data_dict)
-        column, row = convert_cell_to_coordinates(cell)
-        x = data_dict["x"][0]
-        y = data_dict["y"][0]
-        ztotal = data_dict["ztotal"]
-        xfit, yfit = get_fitting_points(x, y, ztotal)
-
-        axs[column].plot(
-            xfit,
-            yfit,
-            label=f"{cell}",
-            color=colors[column],
-            marker=markers[row],
-            markeredgecolor="k",
-            markeredgewidth=0.1,
-        )
-        # plot_optimal_enable_currents(axs[column], data_dict)
-
-        axs[column].legend(loc="upper right", fontsize=6, labelspacing=0.1, handlelength=0.5)
-        axs[column].set_xlim(0, 600)
-        axs[column].set_ylim(0, 1500)
-    return axs
-
-
-def plot_column(axs, dict_list):
-    colors = CMAP3(np.linspace(0.1, 1, 4))
-    markers = ["o", "s", "D", "^"]
-    for data_dict in dict_list:
-        cell = get_current_cell(data_dict)
-        column, row = convert_cell_to_coordinates(cell)
-        x = data_dict["x"][0]
-        y = data_dict["y"][0]
-        ztotal = data_dict["ztotal"]
-        xfit, yfit = get_fitting_points(x, y, ztotal)
-
-        axs[row].plot(
-            xfit,
-            yfit,
-            label=f"{cell}",
-            color=colors[column],
-            marker=markers[row],
-            markeredgecolor="k",
-            markeredgewidth=0.1,
-        )
-        # plot_optimal_enable_currents(axs[row], data_dict)
-        axs[row].legend(loc="upper right", fontsize=6, labelspacing=0.1, handlelength=0.5)
-        axs[row].set_xlim(0, 600)
-        axs[row].set_ylim(0, 1500)
-    return axs
-
-
-def plot_full_grid(axs, dict_list):
-    plot_grid(axs[1:5, 0:4], dict_list)
-    plot_row(axs[0, 0:4], dict_list)
-    plot_column(axs[1:5, 4], dict_list)
-    axs[0, 4].axis("off")
-    axs[4, 0].set_xlabel("Enable Current [µA]")
-    axs[4, 0].set_ylabel("Critical Current [µA]")
-    return axs
-
-
-def plot_waterfall(ax: Axes3D, dict_list: list[dict]) -> Axes3D:
-    colors = CMAP(np.linspace(0.1, 1, len(dict_list)))
-    verts_list = []
-    zlist = []
-    for i, data_dict in enumerate(dict_list):
-        enable_write_currents = get_enable_current_sweep(data_dict, "write")
-        bit_error_rate = get_bit_error_rate(data_dict)
-        write_current = get_write_current(data_dict)
-        ax.plot(
-            enable_write_currents,
-            bit_error_rate,
-            zs=write_current,
-            zdir="y",
-            color=colors[i],
-            marker=".",
-            markerfacecolor="k",
-            markersize=5,
-            linewidth=2,
-        )
-        zlist.append(write_current)
-        verts = polygon_under_graph(enable_write_currents, bit_error_rate, 0.5)
-        verts_list.append(verts)
-
-    poly = PolyCollection(verts_list, facecolors=colors, alpha=0.6, edgecolors="k")
-    ax.add_collection3d(poly, zs=zlist, zdir="y")
-
-    ax.set_xlabel("$I_{{EW}}$ [µA]", labelpad=10)
-    ax.set_ylabel("$I_W$ [µA]", labelpad=70)
-    ax.set_zlabel("BER", labelpad=10)
-    ax.tick_params(axis="both", which="major", labelsize=12, pad=5)
-
-    ax.xaxis.set_rotate_label(True)
-    ax.yaxis.set_rotate_label(False)
-    ax.zaxis.set_rotate_label(True)
-
-    ax.set_zlim(0, 1)
-    ax.set_zticks([0, 0.5, 1])
-    ax.set_ylim(10, zlist[-1])
-    ax.set_yticks(zlist)
-    ax.xaxis.set_pane_color((1.0, 1.0, 1.0, 0.0))
-    ax.yaxis.set_pane_color((1.0, 1.0, 1.0, 0.0))
-    ax.xaxis.set_major_locator(MultipleLocator(50))
-    ax.yaxis.set_major_locator(MultipleLocator(10))
-    ax.set_box_aspect([0.5, 1, 0.2], zoom=0.8)
-    return ax
-
-
 def filter_nan(x, y):
     mask = np.isnan(y)
     x = x[~mask]
@@ -2676,84 +750,4 @@ def filter_nan(x, y):
     return x, y
 
 
-ALPHA = 0.25
-RETRAP = 1
-WIDTH = 0.33
 
-# if __name__ == "__main__":
-#     from nmem.measurement.functions import calculate_power
-
-#     dict_list = import_directory(
-#         r"C:\Users\ICE\Documents\GitHub\nmem\src\nmem\analysis\enable_write_current_sweep\data"
-#     )
-#     data_dict = dict_list[0]
-#     power = calculate_power(data_dict)
-#     persistent_current = 80
-#     fig, ax = plt.subplots()
-#     # critical_current_zero = get_critical_current_intercept(data_dict)*0.88
-#     critical_current_zero = 1250
-#     temperatures = np.linspace(0, CRITICAL_TEMP, 100)
-
-#     plot_calculated_state_currents(
-#         ax,
-#         temperatures,
-#         CRITICAL_TEMP,
-#         RETRAP,
-#         WIDTH,
-#         ALPHA,
-#         persistent_current,
-#         critical_current_zero,
-#     )
-
-#     plot_calculated_filled_region(
-#         ax,
-#         temperatures,
-#         data_dict,
-#         persistent_current,
-#         CRITICAL_TEMP,
-#         RETRAP,
-#         WIDTH,
-#         ALPHA,
-#         critical_current_zero,
-#     )
-#     ax.set_xlabel("Temperature [K]")
-#     ax.set_ylabel("Current [au]")
-#     ax.grid()
-#     ax.set_xlim(0, CRITICAL_TEMP)
-#     ax.plot([7], [800], marker="x", color="black", markersize=10)
-#     ax.legend(frameon=False, loc="upper left", bbox_to_anchor=(1, 1))
-
-#     data_dict1 = sio.loadmat("measured_state_currents_290.mat")
-#     data_dict2 = sio.loadmat("measured_state_currents_300.mat")
-#     data_dict3 = sio.loadmat("measured_state_currents_310.mat")
-
-#     dict_list = [data_dict1, data_dict2, data_dict3]
-#     fit_results = []
-#     for data_dict in [dict_list[2]]:
-#         temp = data_dict["measured_temperature"].flatten()
-#         state_currents = data_dict["measured_state_currents"]
-#         x_list = []
-#         y_list = []
-#         for i in range(4):
-#             x = temp
-#             y = state_currents[:, i]
-#             x, y = filter_nan(x, y)
-#             ax.plot(x, y, "-o", color=RBCOLORS[i], label=f"State {i}")
-
-#         ax.set_xlim(6, 9)
-#         ax.set_ylim(500, 900)
-#     # ax.set_ylim(0, 1)
-#     plt.show()
-
-
-if __name__ == "__main__":
-    dict_list = import_directory(
-        "/home/omedeiro/nmem/src/nmem/analysis/enable_current_relation_v2/data"
-    )
-
-    fig, axs = plt.subplots(
-        5, 5, figsize=(180 / 25.4, 180 / 25.4), sharex=True, sharey=True
-    )
-    plot_full_grid(axs, dict_list)
-    # plt.savefig("enable_current_relation_full_grid.pdf", bbox_inches="tight")
-    plt.show()
